+++ app/ExternalFiles/Tables.php
@@ -12,7 +12,10 @@ defined( 'ABSPATH' ) || exit;
 
 use easyDirectoryListingForWordPress\Directory_Listings;
 use easyDirectoryListingForWordPress\Taxonomy;
+use ExternalFilesInMediaLibrary\ExternalFiles\Protocols\Http;
 use ExternalFilesInMediaLibrary\Plugin\Helper;
+use ExternalFilesInMediaLibrary\Services\Service_Base;
+use ExternalFilesInMediaLibrary\Services\Services;
 use WP_Query;
 use WP_Term_Query;
 use WP_User;
@@ -246,8 +249,7 @@ class Tables {
 	 * @return array<string,string>
 	 */
 	public function add_media_columns( array $columns ): array {
-		$columns['external_files']        = __( 'External file', 'external-files-in-media-library' );
-		$columns['external_files_source'] = __( 'Source', 'external-files-in-media-library' );
+		$columns['external_files'] = __( 'External file', 'external-files-in-media-library' );
 		return $columns;
 	}
 
@@ -261,43 +263,34 @@ class Tables {
 	 */
 	public function add_media_column_content( string $column_name, int $attachment_id ): void {
 		// get the external object for this file.
-		$external_file = Files::get_instance()->get_file( $attachment_id );
+		$external_file_obj = Files::get_instance()->get_file( $attachment_id );
 
 		// show marker if this is an external file.
 		if ( 'external_files' === $column_name ) {
-			// bail if it is not an external file.
-			if ( ! $external_file->is_valid() ) {
-				echo '<span class="dashicons dashicons-no"></span>';
-			} else {
-				echo '<span class="dashicons dashicons-yes"></span>';
+			// show no-icon if it is not an external file.
+			if ( ! $external_file_obj->is_valid() ) {
+				echo '<span class="dashicons dashicons-no" title="' . esc_html__( 'local file', 'external-files-in-media-library' ) .'"></span>';
+				return;
 			}
 
+			// show yes-icon as it is an external file.
+			echo '<span class="dashicons dashicons-yes" title="' . esc_html__( 'external file', 'external-files-in-media-library' ) .'"></span>';
+
 			// show deprecated hint for old hook.
-			do_action_deprecated( 'eml_table_column_content', array( $attachment_id ), '5.0.0', 'efml_table_column_content' );
-
-			/**
-			 * Run additional tasks for show more infos here.
-			 *
-			 * @since 1.0.0 Available since 1.0.0.
-			 * @param int $attachment_id The ID of the attachment.
-			 */
-			do_action( 'efml_table_column_content', $attachment_id );
-		}
-
-		// show additional infos about external files.
-		if ( 'external_files_source' === $column_name && $external_file->is_valid() ) {
-			// get the unproxied URL.
-			$url = $external_file->get_url( true );
+			do_action_deprecated( 'eml_table_column_content', array( $attachment_id ), '5.0.0' );
 
 			// get protocol handler.
-			$protocol_handler = $external_file->get_protocol_handler_obj();
+			$protocol_handler = $external_file_obj->get_protocol_handler_obj();
 
 			// bail if handler could not be found.
 			if ( ! $protocol_handler instanceof Protocol_Base ) {
 				return;
 			}
 
-			// get URL for show depending on used protocol.
+			// get the unproxied URL.
+			$url = $external_file_obj->get_url( true );
+
+			// get URL to show depending on used protocol.
 			$url_to_show = $protocol_handler->get_link();
 
 			// get link or string for the URL.
@@ -307,18 +300,29 @@ class Tables {
 			}
 
 			// get URL.
-			$edit_url = get_edit_post_link( $external_file->get_id() );
+			$edit_url = get_edit_post_link( $external_file_obj->get_id() );
 			if ( ! is_string( $edit_url ) ) {
 				$edit_url = '#';
+			}
+
+			// get the service name used for this file.
+			$service_name = $external_file_obj->get_service_name();
+
+			// get the service object.
+			$service_obj = Services::get_instance()->get_service_by_name( $service_name );
+			$service_title = $protocol_handler->get_title();
+			if ( $service_obj instanceof Service_Base ) {
+				$service_title = $service_obj->get_label();
 			}
 
 			// create dialog.
 			$dialog = array(
 				'title'   => __( 'File info', 'external-files-in-media-library' ),
 				'texts'   => array(
-					'<p><strong>' . __( 'Source', 'external-files-in-media-library' ) . ':</strong> ' . $url_html . '</p>',
-					'<p><strong>' . __( 'Imported at', 'external-files-in-media-library' ) . ':</strong> ' . $external_file->get_date() . '</p>',
-					'<p><strong>' . __( 'Hosting', 'external-files-in-media-library' ) . ':</strong> ' . ( $external_file->is_locally_saved() ? __( 'File is local hosted.', 'external-files-in-media-library' ) : __( 'File is extern hosted.', 'external-files-in-media-library' ) ) . '</p>',
+					'<p><strong>' . __( 'URL', 'external-files-in-media-library' ) . ':</strong> ' . $url_html . '</p>',
+					'<p><strong>' . __( 'Source', 'external-files-in-media-library' ) . ':</strong> ' . $service_title . '</p>',
+					'<p><strong>' . __( 'Imported at', 'external-files-in-media-library' ) . ':</strong> ' . $external_file_obj->get_date() . '</p>',
+					'<p><strong>' . __( 'Hosting', 'external-files-in-media-library' ) . ':</strong> ' . ( $external_file_obj->is_locally_saved() ? __( 'File is local hosted.', 'external-files-in-media-library' ) : __( 'File is extern hosted.', 'external-files-in-media-library' ) ) . '</p>',
 				),
 				'buttons' => array(
 					array(
@@ -334,23 +338,17 @@ class Tables {
 				),
 			);
 
-			// show deprecated hint for old hook.
-			$dialog = apply_filters_deprecated( 'eml_table_column_file_source_dialog', array( $dialog, $external_file ), '5.0.0', 'efml_table_column_file_source_dialog' );
-
 			/**
 			 * Filter the dialog for this file info.
 			 *
 			 * @since 5.0.0 Available since 5.0.0.
 			 * @param array<string,mixed> $dialog The dialog.
-			 * @param File $external_file The external file object.
+			 * @param File $external_file_obj The external file object.
 			 */
-			$dialog = apply_filters( 'efml_table_column_file_source_dialog', $dialog, $external_file );
+			$dialog = apply_filters( 'efml_table_column_file_source_dialog', $dialog, $external_file_obj );
 
 			// get the title.
-			$title = $protocol_handler->get_title();
-
-			// show deprecated hint for old hook.
-			$title = apply_filters_deprecated( 'eml_table_column_source_title', array( $title, $attachment_id ), '5.0.0', 'efml_table_column_source_title' );
+			$title = '<span class="efml-icon efml-' . esc_attr( $service_name ) . '" title="' . esc_attr( $service_title ) .'"></span>';
 
 			/**
 			 * Filter the title for show in source column in media table for external files.
@@ -362,7 +360,7 @@ class Tables {
 			$title = apply_filters( 'efml_table_column_source_title', $title, $attachment_id );
 
 			// output.
-			echo wp_kses_post( $title . ' <a href="' . esc_url( $edit_url ) . '" class="dashicons dashicons-info-outline easy-dialog-for-wordpress" data-dialog="' . esc_attr( Helper::get_json( $dialog ) ) . '"></a>' );
+			echo wp_kses_post( $title . ' <a href="' . esc_url( $edit_url ) . '" class="dashicons dashicons-info-outline easy-dialog-for-wordpress" data-dialog="' . esc_attr( Helper::get_json( $dialog ) ) . '" title="' . esc_html__( 'File info', 'external-files-in-media-library' ) .'"></a>' );
 
 			// show deprecated hint for old hook.
 			do_action_deprecated( 'eml_table_column_source', array( $attachment_id ), '5.0.0', 'efml_table_column_source' );