+++ app/Plugin/Schedules/Synchronization.php
@@ -76,6 +76,12 @@ class Synchronization extends Schedules_Base {
 		// get the arguments.
 		$args = $this->get_args();
 
+		// bail if no method is set.
+		if ( empty( $args['method'] ) ) {
+			Log::get_instance()->create( __( 'Synchronization listing object not set.', 'external-files-in-media-library' ), '', 'error' );
+			return;
+		}
+
 		// get the method object by its name.
 		$directory_listing_obj = Services::get_instance()->get_service_by_name( $args['method'] );