+++ app/Services/GoogleDrive/Protocol.php
@@ -29,6 +29,13 @@ use JsonException;
  */
 class Protocol extends Protocol_Base {
 	/**
+	 * The internal protocol name.
+	 *
+	 * @var string
+	 */
+	protected string $name = 'google-drive';
+
+	/**
 	 * Return whether the file using this protocol is available.
 	 *
 	 * @return bool