+++ changelog.md
@@ -16,6 +16,7 @@
 - Added option to export each newly uploaded file in media library to external sources which are reachable via web
 -> supported for Dropbox, FTP, Google Drive, Google Cloud Storage, local, AWS S3 and WebDav
 -> optionally, you can delete the local files, thereby outsourcing (offloading) all your files and saving storage space
+- Added option to export each file in media library to external source as described above
 - Introduced file handling extensions and added 3 of them (date, queue, real_import)
 - Added option to use the date of external files in add-dialog (2nd file handling extension)
 - Added option to really import files in media library (this disables all external files functions for these files)
@@ -102,6 +103,7 @@
 - Settings for most services are now saved on user and not global, but can be set to global
 - External sources can now get an individual name
 - ZIP files can not also be opened via any supported TCP protocol
+- Save used service on each external file
 - Optimized URL shortener
 
 ### Fixed